+++ src/DetailsView/components/details-view-command-bar.tsx
@@ -11,7 +11,7 @@ import { AssessmentStoreData } from '../../common/types/store-data/assessment-re
 import { FeatureFlagStoreData } from '../../common/types/store-data/feature-flag-store-data';
 import { TabStoreData } from '../../common/types/store-data/tab-store-data';
 import { DetailsViewActionMessageCreator } from '../actions/details-view-action-message-creator';
-import { ReportGeneratorV1, ReportGeneratorDeps } from '../reports/report-generator';
+import { ReportGeneratorDeps, ReportGeneratorV1 } from '../reports/report-generator-v1';
 import { DetailsRightPanelConfiguration } from './details-view-right-panel';
 import { ExportDialog, ExportDialogDeps } from './export-dialog';
 import { StartOverDropdown } from './start-over-dropdown';
