+++ src/DetailsView/reports/report-generator-v1.ts
