+++ src/tests/unit/tests/DetailsView/components/adhoc-issues-test-view.test.tsx
@@ -13,7 +13,7 @@ import { DetailsViewActionMessageCreator } from '../../../../../DetailsView/acti
 import { AdhocIssuesTestView, AdhocIssuesTestViewProps } from '../../../../../DetailsView/components/adhoc-issues-test-view';
 import { IssuesTableHandler } from '../../../../../DetailsView/components/issues-table-handler';
 import { DetailsViewToggleClickHandlerFactory } from '../../../../../DetailsView/handlers/details-view-toggle-click-handler-factory';
-import { ReportGeneratorV1 } from '../../../../../DetailsView/reports/report-generator';
+import { ReportGeneratorV1 } from '../../../../../DetailsView/reports/report-generator-v1';
 import { VisualizationScanResultStoreDataBuilder } from '../../../common/visualization-scan-result-store-data-builder';
 
 describe('AdhocIssuesTestView', () => {
