+++ src/tests/unit/tests/DetailsView/reports/report-generator.test.ts
@@ -6,7 +6,7 @@ import { AssessmentStoreData } from '../../../../../common/types/store-data/asse
 import { FeatureFlagStoreData } from '../../../../../common/types/store-data/feature-flag-store-data';
 import { TabStoreData } from '../../../../../common/types/store-data/tab-store-data';
 import { AssessmentReportHtmlGenerator } from '../../../../../DetailsView/reports/assessment-report-html-generator';
-import { ReportGeneratorV1 } from '../../../../../DetailsView/reports/report-generator';
+import { ReportGeneratorV1 } from '../../../../../DetailsView/reports/report-generator-v1';
 import { ReportHtmlGenerator } from '../../../../../DetailsView/reports/report-html-generator';
 import { ReportNameGenerator } from '../../../../../DetailsView/reports/report-name-generator';
 import { ScanResults } from '../../../../../scanner/iruleresults';
